系统优化

+++ src/main/java/com/zhangheng/file_servser/controller/UpLoadController.java
@@ -1,6 +1,7 @@
 package com.zhangheng.file_servser.controller;
 
 import com.zhangheng.file_servser.entity.Message;
+import com.zhangheng.file_servser.service.KeyService;
 import com.zhangheng.file_servser.service.UpLoadService;
 import com.zhangheng.file_servser.utils.FiletypeUtil;
 import com.zhangheng.file_servser.utils.TimeUtil;
@@ -14,8 +15,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.util.Date;
+import java.util.List;
 import java.util.UUID;
 
 /**
@@ -31,6 +34,10 @@ public class UpLoadController {
 
     @Autowired
     private UpLoadService upLoadService;
+    @Autowired
+    private KeyService keyService;
+    @Value("#{'${admin_keys}'.split(',')}")
+    private List<String> admin_keys;
     @Value("${baseDir}")
     private String baseDir;
     @Value("${appName}")
@@ -152,7 +159,6 @@ public class UpLoadController {
                 msg.setCode(200);
                 msg.setTitle("文件保存成功");
                 msg.setMessage(s);
-                log.info(s);
             }else {
                 msg.setCode(500);
                 msg.setTitle("文件保存失败");
@@ -169,37 +175,49 @@ public class UpLoadController {
 
     @ResponseBody
     @RequestMapping("deleteFile")
-    public Message deleteFile(String path){
+    public Message deleteFile(String path, String key){
         Message msg = new Message();
         msg.setTime(TimeUtil.time(new Date()));
-        if (path!=null&&path.length()>0){
-            File file = new File(baseDir+path);
-            if (file.exists()) {
-                if (!file.isDirectory()) {
-                    boolean b = upLoadService.deleteFile(path);
-                    if (b) {
-                        msg.setCode(200);
-                        msg.setTitle("删除成功");
-                        msg.setMessage("成功！文件删除成功：" + path);
+        if (key!=null&&key.trim().length()>0) {
+            if (keyService.isAdminKeys(key)) {
+                if (path != null && path.length() > 0) {
+                    File file = new File(baseDir + path);
+                    if (file.exists()) {
+                        if (!file.isDirectory()) {
+                            boolean b = upLoadService.deleteFile(path);
+                            if (b) {
+                                msg.setCode(200);
+                                msg.setTitle("删除成功");
+                                msg.setMessage("成功！文件删除成功：" + path);
+                            } else {
+                                msg.setCode(500);
+                                msg.setTitle("删除失败");
+                                msg.setMessage("错误！文件删除失败：" + path);
+                            }
+                        } else {
+                            msg.setCode(404);
+                            msg.setTitle("路径错误");
+                            msg.setMessage("警告！路径：" + path + "不是文件类型");
+                        }
                     } else {
-                        msg.setCode(500);
-                        msg.setTitle("删除失败");
-                        msg.setMessage("错误！文件删除失败：" + path);
+                        msg.setCode(404);
+                        msg.setTitle("文件不存在");
+                        msg.setMessage("警告！删除文件的不存在：" + path);
                     }
-                }else {
-                    msg.setCode(404);
-                    msg.setTitle("路径错误");
-                    msg.setMessage("警告！路径："+path+"不是文件类型");
+                } else {
+                    msg.setCode(500);
+                    msg.setTitle("路径为空");
+                    msg.setMessage("错误！删除路径为空");
                 }
             }else {
-                msg.setCode(404);
-                msg.setTitle("文件不存在");
-                msg.setMessage("警告！删除文件的不存在："+path);
+                msg.setCode(500);
+                msg.setTitle("秘钥错误");
+                msg.setMessage("秘钥错误！请使用管理秘钥操作");
             }
         }else {
             msg.setCode(500);
-            msg.setTitle("路径为空");
-            msg.setMessage("错误！删除路径为空");
+            msg.setTitle("秘钥为空");
+            msg.setMessage("错误！管理秘钥不能为空");
         }
         return msg;
     }
